fix: Ensure image is uint8 before color conversion.

+++ smude/__init__.py
@@ -74,8 +74,9 @@ def enhance_local_contrast_filter(image, radius, **kwargs):
     binary = np.where(contrast_enhanced < threshold, 1, 254)
     binary[mask] = 0
     # Convert back to 3-channel RGB so downstream code expects RGB
+    binary = binary.astype(np.uint8)
     binary = cv2.cvtColor(binary, cv2.COLOR_GRAY2RGB)
-    return binary.astype(np.uint8)
+    return binary
 # -----------------------------------------------------------------------
 
 # Initialize logging
